backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-multipart==0.0.20
pydantic==2.10.6
pytest==8.3.5
httpx==0.28.1
pypdf==5.4.0
python-docx==1.1.2

